dir improvements: move `jsonstream` from `erigon-lib` to `rpc`
