+++ rpc/handler_test.go
@@ -20,13 +20,13 @@ import (
 	"bytes"
 	"context"
 	"errors"
-	jsoniter "github.com/json-iterator/go"
 	"reflect"
 	"testing"
 
+	jsoniter "github.com/json-iterator/go"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/erigontech/erigon-lib/jsonstream"
+	"github.com/erigontech/erigon/rpc/jsonstream"
 )
 
 func TestHandlerDoesNotDoubleWriteNull(t *testing.T) {
