+++ rpc/http.go
@@ -38,10 +38,9 @@ import (
 	"github.com/golang-jwt/jwt/v4"
 
 	"github.com/erigontech/erigon-lib/common"
-	"github.com/erigontech/erigon-lib/jsonstream"
-	"github.com/erigontech/erigon-lib/log/v3"
-
 	"github.com/erigontech/erigon-lib/common/dbg"
+	"github.com/erigontech/erigon-lib/log/v3"
+	"github.com/erigontech/erigon/rpc/jsonstream"
 )
 
 const (
