+++ rpc/jsonrpc/trace_api.go
@@ -22,11 +22,11 @@ import (
 
 	"github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/common/hexutil"
-	"github.com/erigontech/erigon-lib/jsonstream"
 	"github.com/erigontech/erigon/cmd/rpcdaemon/cli/httpcfg"
 	"github.com/erigontech/erigon/db/kv"
 	"github.com/erigontech/erigon/eth/tracers/config"
 	"github.com/erigontech/erigon/rpc"
+	"github.com/erigontech/erigon/rpc/jsonstream"
 )
 
 // TraceAPI RPC interface into tracing API
