+++ rpc/jsonstream/factory.go
