+++ rpc/jsonstream/iterator_stream.go
