+++ rpc/jsonstream/stack_stream.go
