+++ rpc/jsonstream/stack_stream_test.go
