+++ rpc/jsonstream/stream.go
