+++ rpc/jsonstream/stream_benchmark_test.go
